chore(clerk-js): Move all related billing files inside an isolated folder in common
