+++ packages/clerk-js/src/ui/common/index.ts
@@ -16,6 +16,4 @@ export * from './PrintableComponent';
 export * from './NotificationCountBadge';
 export * from './RemoveResourceForm';
 export * from './QRCode';
-export * from './CurrentPlanSection';
-export * from './PaymentMethodSection';
-export * from './Billing';
+export * from './Billing/Billing';
